Get correct default branch
The API does not provide this information, but we don't really need it.

In ideal word we could just do `git ls-remote --symref GIT_URL HEAD` and
get the information, but for some reason this does not currently work
with Pagure.

Default branch should be whatever HEAD points to in the remote repo, and
since we most likely work in a clone of it, the ref will exist locally
as well.